Move +big test into dedicated file
