+++ src/dbnode/storage/database_test.go
@@ -1,5 +1,3 @@
-// +build big
-//
 // Copyright (c) 2016 Uber Technologies, Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -1172,123 +1170,6 @@ func TestDatabaseIsBootstrapped(t *testing.T) {
 
 	assert.True(t, d.IsBootstrapped())
 	assert.False(t, d.IsBootstrapped())
-}
-
-func TestDatabaseIsBootstrappedAndDurable(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	var (
-		validIsBootstrapped                  = true
-		validShardSetAssignedAt              = time.Now()
-		validLastBootstrapCompletionTime     = validShardSetAssignedAt.Add(time.Second)
-		validLastSuccessfulSnapshotStartTime = validLastBootstrapCompletionTime.Add(time.Second)
-	)
-	testCases := []struct {
-		title                           string
-		isBootstrapped                  bool
-		lastBootstrapCompletionTime     time.Time
-		lastSuccessfulSnapshotStartTime time.Time
-		shardSetAssignedAt              time.Time
-		expectedResult                  bool
-	}{
-		{
-			title:                           "False is not bootstrapped",
-			isBootstrapped:                  false,
-			lastBootstrapCompletionTime:     validLastBootstrapCompletionTime,
-			lastSuccessfulSnapshotStartTime: validLastSuccessfulSnapshotStartTime,
-			shardSetAssignedAt:              validShardSetAssignedAt,
-			expectedResult:                  false,
-		},
-		{
-			title:                           "False if no last bootstrap completion time",
-			isBootstrapped:                  validIsBootstrapped,
-			lastBootstrapCompletionTime:     time.Time{},
-			lastSuccessfulSnapshotStartTime: validLastSuccessfulSnapshotStartTime,
-			shardSetAssignedAt:              validShardSetAssignedAt,
-			expectedResult:                  false,
-		},
-		{
-			title:                           "False if no last successful snapshot start time",
-			isBootstrapped:                  validIsBootstrapped,
-			lastBootstrapCompletionTime:     validLastBootstrapCompletionTime,
-			lastSuccessfulSnapshotStartTime: time.Time{},
-			shardSetAssignedAt:              validShardSetAssignedAt,
-			expectedResult:                  false,
-		},
-		{
-			title:                           "False if last snapshot start is not after last bootstrap completion time",
-			isBootstrapped:                  validIsBootstrapped,
-			lastBootstrapCompletionTime:     validLastBootstrapCompletionTime,
-			lastSuccessfulSnapshotStartTime: validLastBootstrapCompletionTime,
-			shardSetAssignedAt:              validShardSetAssignedAt,
-			expectedResult:                  false,
-		},
-		{
-			title:                           "False if last bootstrap completion time is not after shardset assigned at time",
-			isBootstrapped:                  validIsBootstrapped,
-			lastBootstrapCompletionTime:     validLastBootstrapCompletionTime,
-			lastSuccessfulSnapshotStartTime: validLastBootstrapCompletionTime,
-			shardSetAssignedAt:              validLastBootstrapCompletionTime,
-			expectedResult:                  false,
-		},
-		{
-			title:                           "False if last bootstrap completion time is not after/equal shardset assigned at time",
-			isBootstrapped:                  validIsBootstrapped,
-			lastBootstrapCompletionTime:     validLastBootstrapCompletionTime,
-			lastSuccessfulSnapshotStartTime: validLastSuccessfulSnapshotStartTime,
-			shardSetAssignedAt:              validLastBootstrapCompletionTime.Add(time.Second),
-			expectedResult:                  false,
-		},
-		{
-			title:                           "True if all conditions are met",
-			isBootstrapped:                  validIsBootstrapped,
-			lastBootstrapCompletionTime:     validLastBootstrapCompletionTime,
-			lastSuccessfulSnapshotStartTime: validLastSuccessfulSnapshotStartTime,
-			shardSetAssignedAt:              validShardSetAssignedAt,
-			expectedResult:                  true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.title, func(t *testing.T) {
-			d, mapCh, _ := defaultTestDatabase(t, ctrl, Bootstrapped)
-			defer func() {
-				close(mapCh)
-			}()
-
-			mediator := NewMockdatabaseMediator(ctrl)
-			d.mediator = mediator
-			d.lastReceivedNewShards = tc.shardSetAssignedAt
-
-			mediator.EXPECT().IsBootstrapped().Return(tc.isBootstrapped)
-			if !tc.isBootstrapped {
-				assert.Equal(t, tc.expectedResult, d.IsBootstrappedAndDurable())
-				// Early return because other mock calls will not get called.
-				return
-			}
-
-			if tc.lastBootstrapCompletionTime.IsZero() {
-				mediator.EXPECT().LastBootstrapCompletionTime().Return(time.Time{}, false)
-				assert.Equal(t, tc.expectedResult, d.IsBootstrappedAndDurable())
-				// Early return because other mock calls will not get called.
-				return
-			}
-
-			mediator.EXPECT().LastBootstrapCompletionTime().Return(tc.lastBootstrapCompletionTime, true)
-
-			if tc.lastSuccessfulSnapshotStartTime.IsZero() {
-				mediator.EXPECT().LastSuccessfulSnapshotStartTime().Return(time.Time{}, false)
-				assert.Equal(t, tc.expectedResult, d.IsBootstrappedAndDurable())
-				// Early return because other mock calls will not get called.
-				return
-			}
-
-			mediator.EXPECT().LastSuccessfulSnapshotStartTime().Return(tc.lastSuccessfulSnapshotStartTime, true)
-
-			assert.Equal(t, tc.expectedResult, d.IsBootstrappedAndDurable())
-		})
-	}
 }
 
 func TestUpdateBatchWriterBasedOnShardResults(t *testing.T) {
